backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
kagglehub==0.3.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
python-multipart==0.0.19
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
